# application/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "API for managing a mechanic shop, including customers, employees, service tickets, inventory, and services"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions: 
  bearerAuth: # name of the security scheme
    type: apiKey
    name: Authorization
    in: header
    description: "JWT token, prefixed with 'Bearer '"
parameters:
  PageParam:
    name: page
    in: query
    type: integer
    required: false
    default: 1
  LimitParam:
    name: limit
    in: query
    type: integer
    required: false
    default: 10
  SortByParam:
    name: sort_by
    in: query
    type: string
    required: false
  SortOrderParam:
    name: sort_order
    in: query
    type: string
    enum: [asc, desc]
    required: false
  StatusParam:
    name: status
    in: query
    type: string
    enum: [open, in_progress, closed]
    required: false
...
paths:
...
  /customers/me/tickets:
    get:
      tags:
        - Customer
      summary: "Get current customer's service tickets"
      description: "Returns the authenticated customer's service tickets with pagination and optional status filtering."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/StatusParam"
      responses:
        200:
          description: "Tickets retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
              message:
                type: string
              data:
                type: array
                items:
                  $ref: "#/definitions/CreateServiceTicketResponse"
              meta:
                $ref: "#/definitions/PaginationMeta"
        400:
          description: "Validation error"
          schema:
            $ref: "#/definitions/ValidationErrorResponse"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
...
  /employees/me/tickets:
    get:
      tags:
        - Employee
      summary: "Get current employee's service tickets"
      description: "Returns service tickets assigned to the authenticated employee with pagination, sorting, and filtering options. Rate limited to 10 requests per minute."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/StatusParam"
      responses:
        200:
          description: "Service tickets retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              meta:
                $ref: "#/definitions/PaginationMeta"
              data:
                type: array
                items:
                  $ref: "#/definitions/CreateServiceTicketResponse"
          examples:
            application/json:
              status: "success"
              message: "Operation successful"
              meta:
                pagination:
                  has_next: false
                  has_prev: false
                  limit: 10
                  page: 1
                  total_items: 3
                  total_pages: 1
              data:
                - id: 2
                  closed_at: "2025-04-20T00:45:23.330793"
                  cost: 116.58
                  created_at: "2025-04-20T00:45:23.344572"
                  customer:
                    email: "zamzam@gmail.com"
                    id: 1
                    name: "zambak"
                  employees:
                    - id: 1
                      name: "mike"
                      role: "mechanic"
                  is_deleted: false
                  serialized_parts:
                    - id: 1
                      inventory:
                        desc: "Front brake pad set"
                        id: 3
                        inventory_number: "BP-123"
                        name: "Brake Pad"
                        price: "59.99"
                      serial_number: "SP-001"
                      status: "defective"
                  services:
                    - base_price: "49.99"
                      id: 1
                      service_type: "Brake Check"
                  status: "closed"
                  vin: "1HGCM82633A004352"
                  work_summary: "Brake replacement and oil change"
        401:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/UnauthorizedResponse"
        403:
          description: "Forbidden"
          schema:
            $ref: "#/definitions/ForbiddenResponse"
        404:
          description: "Employee not found"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "error"
              message:
                type: string
                example: "Employee not found"
          examples:
            application/json:
              status: "error"
              message: "Employee not found"
        429:
          description: "Too Many Requests"
          produces:
            - text/html
          schema:
            type: string
            example: "<!doctype html>\n<html lang=en>\n<title>429 Too Many Requests</title>\n<h1>Too Many Requests</h1>\n<p>10 per 1 minute</p>"
          examples:
            text/html: |
              <!doctype html>
              <html lang=en>
              <title>429 Too Many Requests</title>
              <h1>Too Many Requests</h1>
              <p>10 per 1 minute</p>
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/ServerErrorResponse"
...
  /services:
    post:
      tags:
        - Service
      summary: "Create new service"
      description: "Creates a new service with the provided details. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateServicePayload"
      responses:
        201:
          description: "Service created successfully"
          schema:
            $ref: "#/definitions/CreateServiceResponse"
        
    get:
      tags:
        - Service
      summary: "Get all services"
      description: "Retrieves all services. Requires employee token."
      responses:
        200:
          description: "Services retrieved successfully"
          schema:
            $ref: "#/definitions/GetServicesResponse"
  /services/{service_id}:
    get:
      tags:
        - Service
      summary: "Get service by ID"
      description: "Retrieves a single service by its ID. Requires employee token."
      parameters:
        - name: service_id
          in: path
          required: true
          type: integer
          description: "ID of the service to retrieve"
      responses:
        200:
          description: "Service retrieved successfully"
          schema:
            $ref: "#/definitions/CreateServiceResponse"

    put:
      tags:
        - Service
      summary: "Update service"
      description: "Updates a service with the provided details. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: service_id
          in: path
          required: true
          type: integer
          description: "ID of the service to update"
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateServicePayload"
      responses:
        200:
          description: "Service updated successfully"
          schema:
            $ref: "#/definitions/CreateServiceResponse"

    patch:
      tags:
        - Service
      summary: "Partially update service"
      description: "Updates specific fields of a service. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: service_id
          in: path
          required: true
          type: integer
          description: "ID of the service to update"
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/PatchServicePayload"
      responses:
        200:
          description: "Service updated successfully"
          schema:
            $ref: "#/definitions/CreateServiceResponse"

    delete:
      tags:
        - Service
      summary: "Delete service"
      description: "Permanently deletes a service. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: service_id
          in: path
          required: true
          type: integer
          description: "ID of the service to delete"
      responses:
        204:
          description: "Service deleted successfully"

#MARK: service ticket paths
  /service-tickets:
    post:
      tags:
        - ServiceTicket
      summary: "Create new service ticket"
      description: "Creates a new service ticket with the provided details. Requires employee token."
      security:
        - bearerAuth: []
      parameters: 
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/CreateServiceTicketPayload"
      responses:  
        201:
          description: "Service ticket created successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Service ticket created successfully"
              data:
                $ref: "#/definitions/CreateServiceTicketResponse"

    get:
      tags:
        - ServiceTicket
      summary: "List all service tickets"
      description: "Returns a paginated list of active service tickets with optional filtering and sorting. Only tickets with is_deleted=False are shown. Cached for 60 seconds."
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/parameters/PageParam"
        - $ref: "#/parameters/LimitParam"
        - $ref: "#/parameters/SortByParam"
        - $ref: "#/parameters/SortOrderParam"
        - $ref: "#/parameters/StatusParam"
        - name: status
          in: query
          type: string
          required: false
          description: "Filter by status"
        - name: customer_id
          in: query
          type: integer
          required: false
          description: "Filter tickets by customer ID"
      responses:
        200:
          description: "Service tickets retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Service tickets retrieved successfully"
              data:
                type: array
                items:
                  $ref: "#/definitions/CreateServiceTicketResponse"
              meta:
                $ref: "#/definitions/PaginationMeta"
  /service-tickets/{service_ticket_id}:
    get:
      tags:
        - ServiceTicket
      summary: "Get service ticket by ID"
      description: "Retrieves a single service ticket by its ID. Requires employee token."
      security: 
        - bearerAuth: []
      parameters:
        - name: service_ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the service ticket to retrieve"
      responses:
        200:
          description: "Service ticket retrieved successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                $ref: "#/definitions/CreateServiceTicketResponse"
      
    patch:
      tags:
        - ServiceTicket
      summary: "Partially update service ticket"
      description: "Updates specific fields of a service ticket. Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: service_ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the service ticket to update"
        - in: body
          name: body
          required: true
          schema:
            $ref: "#/definitions/PatchServiceTicketPayload"
      responses:
        200:
          description: "Service ticket updated successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Operation successful"
              data:
                $ref: "#/definitions/CreateServiceTicketResponse"

    delete:
      tags:
        - ServiceTicket
      summary: "Soft delete a service ticket"
      description: "Marks a service ticket as deleted (soft delete). Requires employee token."
      security:
        - bearerAuth: []
      parameters:
        - name: service_ticket_id
          in: path
          required: true
          type: integer
          description: "ID of the service ticket to delete"
      responses:
        200:
          description: "Service ticket soft-deleted successfully"
          schema:
            type: object
            properties:
              status:
                type: string
                example: "success"
              message:
                type: string
                example: "Service ticket soft-deleted"

#MARK: DEFINITIONS
definitions:
...
  ValidationErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Invalid input"
      errors:
        type: object
        properties:
          field_name:
            type: array
            items:
              type: string
            example: ["Missing data for required field."]
    examples:
      application/json:
        status: "error"
        message: "Invalid input"
  UnauthorizedResponse:
    type: object
    properties:
      message:
        type: string
        example: "Authorization header is required"
    examples:
      application/json:
        message: "Authorization header is required"
  ForbiddenResponse:
    type: object
    properties:
      message:
        type: string
        example: "Forbidden - invalid role"
    examples:
      application/json:
        message: "Forbidden - invalid role"
...
  ServerErrorResponse:
    type: object
    properties:
      status:
        type: string
        example: "error"
      message:
        type: string
        example: "Internal server error"
    examples:
      application/json:
        status: "error"
        message: "Internal server error"
...
  PaginationMeta:
    type: object
    properties:
      pagination:
        type: object
        properties:
          has_next:
            type: boolean
            example: false
          has_prev:
            type: boolean
            example: false
          limit:
            type: integer
            example: 10
          page:
            type: integer
            example: 1
          total_items:
            type: integer
            example: 1
          total_pages:
            type: integer
            example: 1
...
  CreateServicePayload:
    type: object
    required:
      - service_type
      - base_price
      - description 
    properties:
      service_type:
        type: string
        example: "Wheel Alignment"
      base_price:
        type: string
        example: "89.99"  
      description:
        type: string
        example: "Tire Services"
  PatchServicePayload:
    type: object
    properties:
      service_type:
        type: string
        example: "Brake Check"
      base_price:
        type: string
        example: "59.99"
      description:
        type: string
        example: "Full brake inspection"
  CreateServiceResponse:
    type: object
    properties:
      id:
        type: integer
        example: 5
      service_type:
        type: string
        example: "Wheel Alignment"
      base_price:
        type: string
        example: "89.99"
      description:
        type: string
        example: "Tire Services"
  GetServicesResponse:
    type: array
    items:
      $ref: "#/definitions/CreateServiceResponse"

#MARK: service ticket definitions
  CreateServiceTicketPayload:
    type: object
    required:
      - vin
      - work_summary
      - status
      - customer_id
      - employee_ids
      - service_ids
      - part_ids
    properties:
      vin:
        type: string
        example: "ABC9876543210"
        description: "Vehicle Identification Number"
      work_summary:
        type: string
        example: "Brake inspection and pad replacement"
        description: "Summary of work to be performed"
      status:
        type: string
        enum: [open, in_progress, closed]
        example: "open"
      customer_id:
        type: integer
        example: 1
        description: "ID of the customer who owns the vehicle"
      employee_ids:
        type: array
        items:
          type: integer
        example: [2, 3]
        description: "IDs of employees assigned to the ticket"
      service_ids:
        type: array
        items:
          type: integer
        example: [3]
        description: "IDs of services to be performed"
      part_ids:
        type: array
        items:
          type: integer
        example: [3]
        description: "IDs of parts to be used in the service"
  CreateServiceTicketResponse:
    type: object
    properties:
      id:
        type: integer
        example: 8
      vin:
        type: string
        example: "ABC9876543210"
      work_summary:
        type: string
        example: "Brake inspection and pad replacement"
      status:
        type: string
        enum: [open, in_progress, closed]
        example: "open"
      created_at:
        type: string
        format: date-time
        example: "2025-05-21T00:12:03.807165"
      closed_at:
        type: string
        format: date-time
        nullable: true
        example: null
      cost:
        type: number
        format: float
        example: 0.0
      is_deleted:
        type: boolean
        example: false
      customer:
        type: object
        properties:
          id:
            type: integer
            example: 1
          name:
            type: string
            example: "zambak"
          email:
            type: string
            example: "zamzam@gmail.com"
      employees:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 3
            name:
              type: string
              example: "lilo"
            role:
              type: string
              example: "manager"
      services:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 3
            service_type:
              type: string
              example: "Oil Change"
            base_price:
              type: string
              example: "19.99"
      serialized_parts:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
              example: 3
            serial_number:
              type: string
              example: "BT-3501"
            status:
              type: string
              enum: [available, used, defective]
              example: "available"
            inventory:
              type: object
              properties:
                id:
                  type: integer
                  example: 5
                name:
                  type: string
                  example: "Spark Plug"
                inventory_number:
                  type: string
                  example: "SP-200"
                price:
                  type: string
                  example: "14.99"
                desc:
                  type: string
                  example: "Spark plug forengines"
  PatchServiceTicketPayload:
    type: object
    properties:
      work_summary:
        type: string
        example: "Brake inspection and pad replacement"
      status:
        type: string
        enum: [open, in_progress, closed]
        example: "open"
      add_employee_ids:
        type: array
        items:
          type: integer
        example: [1, 2]
        description: "IDs of employees to add to the ticket"
      remove_employee_ids:
        type: array
        items:
          type: integer
        example: [3]
        description: "IDs of employees to remove from the ticket"
      add_service_ids:
        type: array
        items:
          type: integer
        example: [1]
        description: "IDs of services to add to the ticket"
      remove_service_ids:
        type: array
        items:
          type: integer
        example: [2]
        description: "IDs of services to remove from the ticket" 
      add_part_ids:
        type: array
        items:
          type: integer
        example: [1]
        description: "IDs of parts to add to the ticket (must be available parts)"
      remove_part_ids:
        type: array
        items:
          type: integer
        example: [2]
        description: "IDs of parts to remove from the ticket"